requirements.txt
pandasai
pandasai-docker